CalculateAverage returned type changed from int to double

+++ apbd-cw1-git-s33470/StatisticsHelper.cs
@@ -17,9 +17,9 @@ public class StatisticsHelper
         return Convert.ToInt32(hex, 16);
     }
 
-    public static int CalculateAverage(int[] numbers)
+    public static double CalculateAverage(int[] numbers)
     {
-        int sum = 0;
+        double sum = 0;
         foreach (int number in numbers)
         {
             sum += number;
